Updated several components to TypeScript
