+++ tsconfig.json
@@ -1,4 +1,6 @@
 {
+    "include": ["src/**/*"],
+    "exclude": ["node_modules"],
     "compilerOptions": {
         "jsx": "react-jsx",
         /* Visit https://aka.ms/tsconfig to read more about this file */
